requirements.txt
numpy
PyPDF2
pdf2image
pillow
eel
pyinstaller